Adding a ParseMultipartForm line for force ValidateForm to wait until the entire form is ready
I was struggling to upload larger forms, with a large amount of images with large amounts of pre-labels
ValidateForms would query information to validate, before the entire form had been uploaded and parsed, which was leading to a race condition.
This way, the server will wait until the entire form has been uploaded, before attempting to validate it.

100 mB of the forms will be held in memory before the server will start writing chunks to disk. I do not know if this is the right amount for this project, but it worked for me.